@@ -2,8 +2,8 @@ package com.mrgao.onemonth
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.LinearLayoutManager
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.View
 import com.haibin.calendarview.Calendar
 import com.haibin.calendarview.CalendarView
