@@ -2,7 +2,7 @@ package com.mrgao.onemonth
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.support.v7.widget.LinearLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.View
 import com.mrgao.onemonth.base.BaseDatabindingFragment
 import com.mrgao.onemonth.base.BindingAdapter
