@@ -3,7 +3,7 @@ package com.mrgao.onemonth.utils;
 import android.database.Cursor;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 
 import org.greenrobot.greendao.AbstractDao;
